myproject/requirements.txt
bleach==2.1.1
boto==2.48.0
bz2file==0.98
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.1
coreschema==0.0.4
decorator==4.1.2
Django==1.8.4
django-crispy-forms==1.6.1
django-datetime-widget==0.9.3
django-rest-swagger==2.1.2
djangorestframework==3.6.3
entrypoints==0.2.3
html5lib==1.0b10
idna==2.5
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
itypes==1.1.0
jedi==0.11.0
Jinja2==2.9.6
JPype1==0.6.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
lxml==4.1.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.0
numpy==1.13.3
openapi-codec==1.3.2
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.3
selenium==3.6.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.11.0
smart-open==1.5.3
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.3
